Main focus: Improved I2C for EEPROM, seems to be more stable now, even under heavy load.
Revert CAT Mode fix partially, since it did not work as expected
when TXing, needs investigation.

+++ mchf-eclipse/hardware/mchf_hw_i2c2.h
@@ -24,9 +24,9 @@ uchar 	mchf_hw_i2c2_WriteRegister(uchar I2CAddr,uchar RegisterAddr, uchar Regist
 
 void Delay(__IO uint32_t);
 
-uint8_t Write_24Cxx(uint32_t, uint8_t, uint8_t);
+uint16_t Write_24Cxx(uint32_t, uint8_t, uint8_t);
 uint16_t Read_24Cxx(uint32_t, uint8_t);
-uint8_t Write_24Cxxseq(uint32_t, uint8_t*, uint16_t, uint8_t);
-uint8_t Read_24Cxxseq(uint32_t, uint8_t*, uint16_t, uint8_t);
+uint16_t Write_24Cxxseq(uint32_t, uint8_t*, uint16_t, uint8_t);
+uint16_t Read_24Cxxseq(uint32_t, uint8_t*, uint16_t, uint8_t);
 
 #endif
